Cosmonstants.jl

This package is an extension of the package PhysicalConstants.jl to include more constants and things useful for my own work. It is not meant to be complete.

It has not been thoroughly tested for BigFloat due to annoying errors stemming from failing assertions in PhysicalConstants.jl.

Arguably, one of the main functionalities is the possibility to decide between Unitfull and Unitless constants. The former behaves exactly like PhysicalConstants.jl, whereas the latter provides only the numerical values of the constants in S.I. units (as Float64). This is very useful for dispatching.

Cosmonstants.referencesDict — Constant

The dictionary of references.

Cosmonstants.correspondingSymbol — Method

Given a constant, return its short name.

Cosmonstants.fullReference — Method

Get the full reference for a given source of information.

Cosmonstants.getCorpusclesParticleMassUncertainty — Method

Helper function (unexported) to get the mass uncertainty value of a particle in kg. If lower and upper uncertainties are different, return the largest one of the two. This is done to be fully consistent with PhysicalConstants.jl, which uses the package Measurements.jl, which does not accept asymmetric uncertainties.

Input

. p:: ParticleType object for Corpuscles.jl

Output

. Particle mass uncertainty (without explicit units) in kg.

Cosmonstants.getCorpusclesParticleMassValue — Method

Helper function (unexported) to get the mass value of a particle in kg.

Input

. p:: ParticleType object for Corpuscles.jl

Output

. Particle mass (without explicit units) in kg.

Cosmonstants.getListOfConstants — Method

Get list of constants. This is useful to remove items that are not constants to create table of constants.

Cosmonstants.@captureShowOutput — Macro

@captureShowOutput(ex)

A macro that captures the output of evaluating an expression ex and returns a tuple containing the result of the evaluation and a string representation of the output.

Input

. ex: the expression to be evaluated (e.g., a constant or a mathematical expression).

Output

A tuple (val, output_string) where:
. val is the result of evaluating the expression ex
. output_string is a string containing the expression, its result, and its textual representation

Cosmonstants.@createConstantsForMasses — Macro

Macro to conveniently create objects like in PhysicalConstants.jl from particle masses contained in Corpuscles.jl.

Input

. particle: a Corpuscles's Particle-type object
. name: explicit name of the variable of the corresponding constant
. symbol: short alias for the corresponding quantity (must be explicitly imported)
. info: string containing the details of the quantity.

To do

. automatically retrieve PDG year (Corpuscles._default_year).

Cosmonstants.Unitfull — Module

module Unitfull

This module defines a set of physical constants and derived quantities. It imports fundamental constants from external modules and provides shorthand notations for convenience. By using the module, the corresponding constants/units are scope-protected.

Cosmonstants.Unitless — Module

module Unitless

This module defines a set of physical constants and derived quantities stripping its units. The default is S.I. units.